+++ src/lib/log.test.ts
@@ -1,20 +1,20 @@
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
 
-import {Logger, type LoggerState} from './log.js';
+import {Logger, type Logger_State} from './log.js';
 
 /* test__Logger */
-interface TestLoggerContext {
-	loggedArgs: any;
-	loggerState: LoggerState;
+interface Test_Logger_Context {
+	logged_args: any;
+	logger_state: Logger_State;
 }
-const createTest__loggerContext = (): TestLoggerContext => {
-	const ctx: TestLoggerContext = {
-		loggedArgs: undefined, // stores the result of the latest log call
-		loggerState: {
+const create_test_logger_context = (): Test_Logger_Context => {
+	const ctx: Test_Logger_Context = {
+		logged_args: undefined, // stores the result of the latest log call
+		logger_state: {
 			level: 'debug',
-			log: (...logArgs: any[]) => {
-				ctx.loggedArgs = logArgs;
+			log: (...log_args: any[]) => {
+				ctx.logged_args = log_args;
 			},
 			prefixes: ['pre'],
 			suffixes: ['post'],
@@ -38,13 +38,13 @@ const createTest__loggerContext = (): TestLoggerContext => {
 	};
 	return ctx;
 };
-const test__Logger = suite('Logger', createTest__loggerContext());
+const test__Logger = suite('Logger', create_test_logger_context());
 
 test__Logger('prefixes and suffixes', (ctx) => {
-	const log = new Logger(['p1', 'p2'], ['s1', 's2'], ctx.loggerState);
+	const log = new Logger(['p1', 'p2'], ['s1', 's2'], ctx.logger_state);
 
 	log.error('foo', 36);
-	assert.equal(ctx.loggedArgs, [
+	assert.equal(ctx.logged_args, [
 		'pre',
 		'errorP1',
 		'errorP2',
@@ -58,10 +58,10 @@ test__Logger('prefixes and suffixes', (ctx) => {
 		'errorS2',
 		'post',
 	]);
-	ctx.loggedArgs = undefined;
+	ctx.logged_args = undefined;
 
 	log.warn('foo', 36);
-	assert.equal(ctx.loggedArgs, [
+	assert.equal(ctx.logged_args, [
 		'pre',
 		'warnP1',
 		'warnP2',
@@ -75,10 +75,10 @@ test__Logger('prefixes and suffixes', (ctx) => {
 		'warnS2',
 		'post',
 	]);
-	ctx.loggedArgs = undefined;
+	ctx.logged_args = undefined;
 
 	log.info('foo', 36);
-	assert.equal(ctx.loggedArgs, [
+	assert.equal(ctx.logged_args, [
 		'pre',
 		'infoP1',
 		'infoP2',
@@ -92,10 +92,10 @@ test__Logger('prefixes and suffixes', (ctx) => {
 		'infoS2',
 		'post',
 	]);
-	ctx.loggedArgs = undefined;
+	ctx.logged_args = undefined;
 
 	log.debug('foo', 36);
-	assert.equal(ctx.loggedArgs, [
+	assert.equal(ctx.logged_args, [
 		'pre',
 		'debugP1',
 		'debugP2',
@@ -109,52 +109,52 @@ test__Logger('prefixes and suffixes', (ctx) => {
 		'debugS2',
 		'post',
 	]);
-	ctx.loggedArgs = undefined;
+	ctx.logged_args = undefined;
 });
 
 test__Logger('mutate logger state to change prefix and suffix', (ctx) => {
 	const log = new Logger(undefined, undefined, {
-		...ctx.loggerState,
+		...ctx.logger_state,
 		info: {
 			prefixes: ['p1', 'p2'],
 			suffixes: ['s1', 's2'],
 		},
 	});
 	log.info('foo', 36);
-	assert.equal(ctx.loggedArgs, ['pre', 'p1', 'p2', 'foo', 36, 's1', 's2', 'post']);
-	ctx.loggedArgs = undefined;
+	assert.equal(ctx.logged_args, ['pre', 'p1', 'p2', 'foo', 36, 's1', 's2', 'post']);
+	ctx.logged_args = undefined;
 
 	// mutate the prefixes and suffixes
 	log.state.info.prefixes.pop();
 	log.state.info.suffixes.shift();
 
 	log.info('foo', 36);
-	assert.equal(ctx.loggedArgs, ['pre', 'p1', 'foo', 36, 's2', 'post']);
-	ctx.loggedArgs = undefined;
+	assert.equal(ctx.logged_args, ['pre', 'p1', 'foo', 36, 's2', 'post']);
+	ctx.logged_args = undefined;
 });
 
 test__Logger('mutate logger state to change log level', (ctx) => {
 	const state = {
-		...ctx.loggerState,
+		...ctx.logger_state,
 		info: {prefixes: [], suffixes: []},
 		warn: {prefixes: [], suffixes: []},
 	};
 	const log = new Logger(undefined, undefined, state);
 
 	log.info('foo');
-	assert.equal(ctx.loggedArgs, ['pre', 'foo', 'post']);
-	ctx.loggedArgs = undefined;
+	assert.equal(ctx.logged_args, ['pre', 'foo', 'post']);
+	ctx.logged_args = undefined;
 
 	state.level = 'warn';
 
 	// `info` should now be silenced
 	log.info('foo');
-	assert.equal(ctx.loggedArgs, undefined);
+	assert.equal(ctx.logged_args, undefined);
 
 	// `warn` is not silenced though
 	log.warn('foo');
-	assert.equal(ctx.loggedArgs, ['pre', 'foo', 'post']);
-	ctx.loggedArgs = undefined;
+	assert.equal(ctx.logged_args, ['pre', 'foo', 'post']);
+	ctx.logged_args = undefined;
 });
 
 test__Logger.run();